multiple scale text output fixing _WIN32 #defines

+++ Test/TimeSeries/milk-tea-coffee.h
@@ -2,9 +2,9 @@
 
 typedef struct yeardatum {
 	int year;
-	float milk;
-	float tea;
-	float coffee;
+	double milk;
+	double tea;
+	double coffee;
 } yeardatum;
 
 yeardatum data[] = {
